requirements.txt
Flask==2.3.3
celery==5.6.5
python-dotenv==1.0.1
gunicorn==21.2.0
uvicorn==0.23.2
 